Ohsung

+++ controllers/comments.controller.js
@@ -15,11 +15,15 @@ class CommentController {
       const { postIdx } = req.params;
       const { desc } = req.body;
       const { userIdx } = res.locals.user;
-      const resultSchema = await commentSchema.validate({ desc });
 
-      // 검증 실패 시 오류 메세지 반환
+      const resultSchema = await commentSchema.validate({ desc });
       if (resultSchema.error && desc.length < 1 ) {
         throw Boom.preconditionFailed('댓글 내용을 입력해주세요.');
+      }
+
+      const seletPost = await this.commentService.findPost(postIdx)
+      if (!seletPost) {
+        throw Boom.notFound('해당 게시글은 존재하지 않습니다.');
       }
 
       await this.commentService.createComment(
@@ -29,7 +33,6 @@ class CommentController {
       );
 
       return res.status(200).json({
-          success: true,
           message: '댓글을 작성하였습니다.',
       });
     } catch (error) {
@@ -43,10 +46,15 @@ class CommentController {
     try {
       const { postIdx } = req.params;
 
-      const selectComments = await this.commentService.selectComments(postIdx);
+      const seletPost = await this.commentService.findPost(postIdx)
+      if (!seletPost) {
+        throw Boom.notFound('해당 게시글은 존재하지 않습니다.');
+      }
+
+      const seletComments = await this.commentService.findComment(postIdx);
 
       return res.status(200).json({
-        comments: selectComments,
+        comments: seletComments,
       });
     } catch (error) {
         logger.error(error.message);
@@ -61,11 +69,15 @@ class CommentController {
       const { commentIdx } = req.params;
       const { desc } = req.body;
       const { userIdx } = res.locals.user;
+
       const resultSchema = await commentSchema.validate({ desc });
-      
-      // 스키마를 이용해 데이터를 검증한다
       if (resultSchema.error && desc.length < 1 ) {
         throw Boom.preconditionFailed('댓글 내용을 입력해주세요.');
+      }
+
+      const chkAuthComment = await this.commentService.findeAuth( commentIdx, userIdx );
+      if (!chkAuthComment) {
+        throw Boom.forbidden('댓글의 수정 권한이 없습니다.');
       }
 
       await this.commentService.updateComment(
@@ -74,59 +86,39 @@ class CommentController {
         userIdx
       );
 
-      
-        return res.status(200).json({
-          message: '댓글을 수정하였습니다.',
-        });
+      return res.status(200).json({
+        message: '댓글을 수정하였습니다.',
+      });
     } catch (err) {
         logger.error(error.message);
         throw error;
     }
   };
 
-  /**
-   * 댓글 삭제
-   */
+
+   //댓글 삭제
   deleteComment = async (req, res, next) => {
     try {
       const { commentIdx } = req.params;
       const { userIdx } = res.locals.user;
 
-      const deleteComment = await this.commentService.deleteComment(
+      const chkAuthComment = await this.commentService.findeAuth( commentIdx, userIdx );
+      if (!chkAuthComment) {
+        throw Boom.forbidden('댓글의 삭제 권한이 없습니다.');
+      }
+
+      await this.commentService.deleteComment(
         commentIdx,
         userIdx
       );
 
-      if (deleteComment === 0) {
-        throw Boom.badRequest('댓글 삭제에 실패했습니다', false);
-      } else {
-        return res.status(200).json({
-          success: true,
-          message: '댓글을 삭제하였습니다.',
-        });
-      }
-    } catch (err) {
-      next(err);
-    }
-  };
-
-  /**
-   * 댓글 권한 확인
-   * @returns res
-   */
-  getAuth = async (req, res, next) => {
-    try {
-      const { commentId } = req.params;
-      const { userId } = res.locals.user;
-      await this.commentService.checkAuth(commentId, userId);
-      // const chk = await this.quizService.checkAuth(quizId, userId)
       return res.status(200).json({
-        success: true,
-        message: '수정 및 삭제 권한이 확인 되었습니다.',
+        message: '댓글을 삭제하였습니다.',
       });
-    } catch (error) {
-      next(error);
-    }
+    } catch (err) {
+      logger.error(error.message);
+      throw error;
+  }
   };
 }
